# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: SimonCloud API
  version: "1.0.0"
  description: |
    API REST de SimonCloud — plataforma documental UMSS.

    **Trazabilidad**: Todos los endpoints están trazados a su FSD-UC correspondiente.
    **Autenticación**: JWT Bearer (emitido por auth-service tras intercambio con WebSISS SSO).
    **Usuario Externo**: Acceso via token temporal firmado como query param `?token=` (sin JWT).
    **Integridad**: Hash SHA-256 de cada archivo, siempre generado en el servidor (nunca en cliente).

    Documentación completa: `docs/fsd/FSD_vFinal.md`

  contact:
    name: Carlos Alberto Gomez Ormachea
    email: beto.sempai@gmail.com
servers:
  - url: http://localhost:3000
    description: Dev local (NestJS + PostgreSQL + MinIO)
  - url: https://simoncloud.umss.edu.bo/api
    description: Producción on-premise DTIC-UMSS
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por auth-service (TTL 8 horas). Almacenado en cookie HttpOnly en el frontend.
  schemas:
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...
    ISO8601DateTime:
      type: string
      format: date-time
      example: "2026-10-15T20:00:00-04:00"
...
    ErrorResponse:
      type: object
      required: [statusCode, message, errorCode, timestamp]
      properties:
        statusCode:
          type: integer
          example: 422
        message:
          type: string
          example: "Regla de negocio violada"
        errorCode:
          type: string
          example: "SIMONDROP_CERRADO"
          description: Código de error trazable a BR-NNN del FSD §5
        timestamp:
          $ref: "#/components/schemas/ISO8601DateTime"

    # Auth schemas
...
    CreateDropRequest:
      type: object
      required: [titulo, cierreEn]
      properties:
        titulo:
          type: string
          minLength: 5
          maxLength: 200
          example: "Entrega Proyecto Final INF-401 2026-2"
        descripcion:
          type: string
          maxLength: 1000
        cierreEn:
          $ref: "#/components/schemas/ISO8601DateTime"
        lmsAssignmentId:
          type: string
          nullable: true
          description: ID de la tarea en Moodle/Classroom si el Drop fue creado via LTI (FSD-UC-001)
        lmsCourseId:
          type: string
          nullable: true
    DropResponse:
      type: object
      required: [id, titulo, cierreEn, status, docenteId, creadoEn]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        titulo:
          type: string
        descripcion:
          type: string
          nullable: true
        cierreEn:
          $ref: "#/components/schemas/ISO8601DateTime"
        status:
          type: string
          enum: [ACTIVO, CERRADO]
        docenteId:
          $ref: "#/components/schemas/UUID"
        lmsAssignmentId:
          type: string
          nullable: true
        totalArchivos:
          type: integer
          minimum: 0
        creadoEn:
          $ref: "#/components/schemas/ISO8601DateTime"
...
security:
  - bearerAuth: []
...
paths:
...
  /drops:
    post:
      tags: [drops]
      operationId: createDrop
      summary: "FSD-UC-001 — Crear SimonDrop"
      description: |
        Crea un buzón de entrega. Requiere rol DOCENTE.
        Si `lmsAssignmentId` está presente, el Drop fue creado via integración LTI desde Moodle/Classroom.
        PRD-US-001, PRD-REQ-001, BR-006.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDropRequest"
      responses:
        "201":
          description: SimonDrop creado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DropResponse"
        "400":
          description: Datos de entrada inválidos (cierreEn en el pasado, título muy corto)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: JWT inválido o expirado
        "403":
          description: "Rol insuficiente — requiere DOCENTE"

    get:
      tags: [drops]
      operationId: listDropsByDocente
      summary: Listar SimonDrops del docente autenticado
      description: Retorna los SimonDrops creados por el docente autenticado. Paginado.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVO, CERRADO]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Lista de SimonDrops
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DropResponse"
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
  /drops/{dropId}:
    get:
      tags: [drops]
      operationId: getDrop
      summary: Obtener detalle de un SimonDrop
      parameters:
        - name: dropId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Detalle del SimonDrop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DropResponse"
        "404":
          description: SimonDrop no encontrado
  /drops/{dropId}/close:
    post:
      tags: [drops]
      operationId: closeDrop
      summary: "FSD-UC-001 — Cerrar SimonDrop manualmente"
      description: Cierra el SimonDrop antes de su fecha de cierre programada. Requiere ser el docente propietario. BR-005.
      parameters:
        - name: dropId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: SimonDrop cerrado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DropResponse"
        "422":
          description: El SimonDrop ya está cerrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...